Generator: Refactor WriterService in multiple services
