+++ Generator/Container.cs
@@ -1,6 +1,8 @@
 ﻿using Generator;
 using Generator.Factories;
 using Generator.Services;
+using Generator.Services.Writer;
+using Generator.Writer;
 using StrongInject;
 
 namespace Repository
@@ -8,6 +10,9 @@ namespace Repository
     [Register(typeof(WriterService))]
     [Register(typeof(TemplateReaderService), typeof(ITemplateReaderService))]
     [Register(typeof(DllImportResolverFactory), typeof(IDllImportResolverFactory))]
+    [Register(typeof(WriteHelperService), typeof(IWriteHelperService))]
+    [Register(typeof(WriteDllImportService), typeof(IWriteDllImportService))]
+    [Register(typeof(WriteTypesService), typeof(IWriteTypesService))]
     public partial class Container : IContainer<WriterService>
     {
     }
